J1708Message and KLineStatusEvent done.

+++ src/Vector/BLF/KLineStatusEvent.h
@@ -72,7 +72,7 @@ public:
     DWORD port;
 
     /** reserved */
-    DWORD reserved;
+    ULONGLONG reserved;
 
     /**
      * @brief the actual data, but only dataLen BYTES are used!
